Update default config to bind all interfaces
The default port configuration was only binding localhost. This works
well in local development, but means that when deployed to Nomad these
ports are only accessible within that allocation.

Binding all interfaces will allow access from other apps running on that
server and with the addition of port forwarding allow requests from
other servers as well.