Fix nomad validation errors and generally improve
The nomad plan fails validation because it is using the outdated schema.
Moving the network and service stanzas to the group level as is the best
practice now to fix the validation errors.

Also, improving the security and deployment times by using the proper
environment variable handling in nomad rather than the custom start
script model. This is the standard we should be adopting across all
applications moving forwards, but have not had the chance to roll out
yet.